rework vacation reply
